% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcw.R
\name{ipcw}
\alias{ipcw}
\title{Computing the stabilized IPCweights}
\usage{
ipcw(
  data,
  id,
  tstart,
  tstop,
  cens,
  arm,
  bas.cov,
  conf,
  trunc = NULL,
  type = "kaplan-meier"
)
}
\arguments{
\item{data}{a dataframe containing the following variables}

\item{id}{the patient's id}

\item{tstart}{the date of the beginning of the follow-up (in numeric format, with the first being equal at 0)}

\item{tstop}{the date of the end of the follow-up (in numeric format)}

\item{cens}{the indicator of treatment censoring (denoted by 1 at the end of the follow-up)}

\item{arm}{the randomized treatment (2-levels factor)}

\item{bas.cov}{a vector the baseline covariates}

\item{conf}{a vector of time-dependent confounders}

\item{trunc}{an optional fraction for the weights. For instance, when trunc = 0.01,
the left tail is truncated to the 1st percentile and the right tail is truncated to the 99th percentile}

\item{type}{a character string specifying the type of survival curve. The default is \code{type=`kaplan-meier`}}
}
\value{
the initial dataframe data with stabilized IPCweights as additional arguments. By default, the un-truncated stabilized weights are given. If the trunc option is not NULL then the truncated stabilized weights are also given.
}
\description{
Computing the stabilized IPCweights
}
\examples{
## Not run
# ipcw(toy.rep, tstart = tstart, tstop = tstop, cens = cens,
# arm="arm",
# bas.cov = c("age"),
# conf = c("TDconf"), trunc = 0.05)

# see ?SHIdat for a complete example
}
\references{
Graffeo, N., Latouche, A., Le Tourneau C., Chevret, S. (2019) "ipcwswitch: an R package for inverse probability of censoring weighting with an application to switches in clinical trials". Computers in biology and medicine, 111, 103339. doi : "10.1016/j.compbiomed.2019.103339"
}
\seealso{
\code{\link{SHIdat}}
}
